Fixed some synchronization issues. Added a equal to sema_up and that fixed hanging when 1 child exits.

+++ pintos/src/vm/page.h
@@ -21,7 +21,7 @@ struct page_entry {
   struct frame *frame;      // Address of physical memory entry
   struct swap_slot *swap;   // Address of swap slot
   struct file* file;        // Address of file
-  uint32_t offset;          // Offset of the page into the file
+  uint64_t offset;          // Offset of the page into the file
   uint32_t read_bytes;      // How many to read from the file starting at offset
 
   struct list_elem elem;    // List element for thread-based page table
